Refactor: 클래스 분리 #55
- HLS를 담당하는 로직은 별도의 클래스로 분리
